+++ ceammc/ext/src/ui/ui_radio.h
@@ -66,6 +66,9 @@ public:
     void m_random(const AtomListView& lv);
     void m_plus(t_float f);
     void m_minus(t_float f);
+    void m_hexbeat(const AtomListView& lv);
+    void m_euclid(const AtomListView& lv);
+    void m_rotate(t_float f);
 
 public:
     static void setup();